services/communication-service/requirements.txt
Django==5.0.1
djangorestframework-simplejwt==5.3.1
django-ninja==1.1.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
psycopg2-binary==2.9.9
python-decouple==3.8
redis==5.0.1
requests==2.31.0


